Put the splits table in the race kit, where every game can reach it
Three of the splits table's call sites were outside flash cards: the
typing results screen and the lesson results screen both reached into
`@/games/flashcards/results/` for it, which made the flash-card game a
dependency of the typing game for one panel that knows nothing about
either.

It never did. It reads a finished `Session` — prompt, answer, what was
given, how long each card took — and prints the rows. That is the race
minus the game, so it now lives in `src/games/race/` and is exported
from its index alongside the clock, the lane and the HUD. All three
call sites import it from `@/games/race`, and no game imports another
game's internals anywhere.

The component itself is unchanged, so the race, typing and lesson
results screens render exactly what they rendered before.

quick-win: sorted the out-of-order `@/engine/random` and
`@/engine/records` imports in TypingResults, matching its two sibling
results screens.

Closes #210
