+++ src/games/race/SplitsTable.tsx
