+++ src/games/typing/LessonResults.tsx
@@ -11,8 +11,7 @@ import { randomSeed } from "@/engine/random";
 import { cumulativeSplits, sessionsFor } from "@/engine/records";
 import { ladderProgress } from "@/engine/typing/ladder";
 import { verdictFor } from "@/engine/typing/verdict";
-import { Rewards } from "@/games/race";
-import { SplitsTable } from "@/games/flashcards/results/SplitsTable";
+import { Rewards, SplitsTable } from "@/games/race";
 import { sfx } from "@/services/sound";
 import { PassBars } from "./PassBars";
 import { lessonConfig } from "./lessonRun";
